// src/app/sub-network/pack/net-rest/interface.ts
/* tslint:disable */


export interface Person {
  systemID: number;
  name: string;
  dbName: string;
  dbUser: string;
  dbPass: string;
  dbPort: string;
  dbServer: string;
  identifier: string;
  systemType: string;
  sourcekey: string;
  pin: string;
  result: string;
  commission: number;
  memberPrice: number;
}

export interface Venue {
  systemID: number;
  name: string;
  dbName: string;
  dbUser: string;

}

// private Integer systemID = 0;
// private String name = "";
//
// private String dbName = "";
// private String dbUser = "";
// private String dbPass = "";
// private String dbPort = "";
// private String dbServer = "";
//
// private String identifier = "";
// private String systemType = "";
//
// private String sourcekey = "";
// private String pin = "";
// private String result = "";
//
// private int commission = 0;
// private int memberPrice = 0;